# realestate/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real Estate API",
        "description": "API documentation for the Real Estate Laravel Application",
        "contact": {
            "email": "support@yourapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/inquiries": {
            "get": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Get all inquiries (admin) or own inquiries (user)",
                "operationId": "0002d5f7ccef92ece21eb4fe4d42d4f7",
                "responses": {
                    "200": {
                        "description": "List of inquiries"
                    },
                    "404": {
                        "description": "No inquiries found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Submit a property inquiry",
                "operationId": "3f728b15f306ab8aecbf3a468e2a4a56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "property_id",
                                    "message"
                                ],
                                "properties": {
                                    "property_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inquiry submitted successfully"
                    },
                    "403": {
                        "description": "Only regular users can create inquiries"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inquiries/{id}/users": {
            "get": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Get all inquiries made by a specific user (admin only)",
                "operationId": "eb3675f3fd6740d5564cb36faef8b51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID whose inquiries are to be retrieved",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of inquiries"
                    },
                    "403": {
                        "description": "Only admin can access this"
                    },
                    "404": {
                        "description": "No inquiries found for the user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inquiries/{id}": {
            "get": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Get a specific inquiry by ID",
                "operationId": "8f67c820771f0052915eb14aa7700dcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inquiry details"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Inquiry not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Delete an inquiry by ID",
                "operationId": "a23abef41dad950355787c27a4038030",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inquiry ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inquiry deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Inquiry not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}